Move MockPlayer and simulate_play to own file, move some tests from test_player.py to mock_test_player.py
